+++ src/main/kotlin/com/example/demo/entity/TransferMoneyRequest.kt
@@ -2,4 +2,4 @@ package com.example.demo.entity
 
 import java.math.BigDecimal
 
-data class TransferMoneyRequest(val accountId: Long, val value: BigDecimal, val targetAccount: Long)
+data class TransferMoneyRequest(val accountId: Int, val value: BigDecimal, val targetAccount: Int)
